requirements.txt
openai
pandas
requests
python-dotenv
langchain 
langchain-openai 
langchain-community 
chromadb 
pypdf 
tiktoken